Get the last two pages of events on page load
When the page loads, we want to retrieve and initially display enough
content for the scrollbar to show. If the very last page doesn't
have enough content for the scrollbar to show, the user won't be able
to scroll up to see more job history. To avoid this scenario, we always
fetch the last _two_ pages when loading the view.